Улучшена безопастность. Устранены баги с сохранением прогресса. Добавлены закладки для PDF файлов. Добавлена возможность удалять закладки. Мелкие изменения дизайна и функциональности.

+++ includes/lib.php
@@ -21,7 +21,7 @@ switch ($_POST['function']){
             $progress=json_decode($item['progress'],true);
             $show.= '<div class="col-md-3 col-lg-3 col-sm-6 col-xs-12">
             <div class="panel panel-default">
-            <div class="panel-heading"><h3 class="panel-title name-book">'.$item['author'].'</h3>
+            <div class="panel-heading"><h3 class="panel-title name-book" id="nameBook'.$item['id'].'">'.$item['author'].'</h3>
             </div>
             <div class="panel-body preview-book"><img  class="preview-book-cover" src="'.$item['cover'].'"/></div>
             <div class="panel-footer"><div class="btn-group">
@@ -30,9 +30,9 @@ switch ($_POST['function']){
         <b class="caret"></b>
     </button>
     <ul class="dropdown-menu" role="menu">
-        <li><a href="'.$item['path'].'" download="'.$item['original_name'].'.'.$file_info['extension'].'">Download</a></li>
+        <li><a href="download?id='.$item['id'].'">Download</a></li>
         <li class="divider"></li>
-        <li><a href="" onClick="deleteBook('.$item['id'].')">Delete</a></li>
+        <li><a class="btn-del" onclick="modalDelete('.$item['id'].')">Delete</a></li>
     </ul>
 </div><span class="progress-info pull-right">'.$progress['progress'].'%</span></div>
             </div>
@@ -44,20 +44,25 @@ switch ($_POST['function']){
         if ($_COOKIE['logged_user']!=null && checkKey($_COOKIE['key'])) {
             $stmt = B::selectFromBase('users_files', null, array('id'), array($_POST['id']));
             $data = $stmt->fetchAll();
-            $s = explode("/", $data[0]['path']);
-            $file_path = str_replace("/" . $s[count($s) - 1], '', $data[0]['path']);
-            $path = $_SERVER['DOCUMENT_ROOT'] . '/' . $file_path;
-            $stmt = B::selectFromBase('users_info', null, array('login'), array($_COOKIE['logged_user']));
-            $info = $stmt->fetchAll();
-            $size = $info[0]['files_disk_space'] - filesize($_SERVER['DOCUMENT_ROOT'] . '/' . $data[0]['path']);
-            $last = json_decode($info[0]['last_books'], true);
-            for ($i = 0; $i < count($last); $i++) {
-                if ($last[$i] == $_POST['id']) $last[$i] = 0;
+            $ud = B::selectFromBase('users', null, array('login'), array($_COOKIE['logged_user']));
+            $udata=$ud->fetchAll();
+            if ($udata[0]['id']==$data[0]['id_user']) {
+                $s = explode("/", $data[0]['path']);
+                $file_path = str_replace("/" . $s[count($s) - 1], '', $data[0]['path']);
+                $path = $_SERVER['DOCUMENT_ROOT'] . '/' . $file_path;
+                $stmt = B::selectFromBase('users_info', null, array('login'), array($_COOKIE['logged_user']));
+                $info = $stmt->fetchAll();
+                $size = $info[0]['files_disk_space'] - filesize($_SERVER['DOCUMENT_ROOT'] . '/' . $data[0]['path']);
+                $last = json_decode($info[0]['last_books'], true);
+                for ($i = 0; $i < count($last); $i++) {
+                    if ($last[$i] == $_POST['id']) $last[$i] = 0;
+                }
+                B::updateBase('users_info', array('files_disk_space', 'last_books'), array($size, json_encode($last)), array('login'), array($_COOKIE['logged_user']));
+                removeDirectory($path);
+                unlink('../'.$data[0]['cover']);
+                B::deleteFromBase('users_files', array('id'), array($data[0]['id']));
+                B::deleteFromBase('bookmarks', array('id_book'), array($data[0]['id']));
             }
-            B::updateBase('users_info', array('files_disk_space', 'last_books'), array($size, json_encode($last)), array('login'), array($_COOKIE['logged_user']));
-            removeDirectory($path);
-            B::deleteFromBase('users_files', array('id'), array($data[0]['id']));
-            B::deleteFromBase('bookmarks',array('id_book'),array($data[0]['id']));
         } else {
             delCookies('logged_user');
             delCookies('key');
